Updated IoT Lab Series
### Built Lab 0 - NiFi DataFlow
     - Explore some of the trucking demo directories
     - Run mvn command to compile, package jar files
     - 8 processors to ingest, route and land data into a specific
location
     - verify that the random data generated for logs and truck events
is separated/routed to their appropriate directories

### Updated Lab 1 - Kafka
     - Replaced Kafka producer with NiFi DataFlow because it sends the
data to a specific location
     - Added PutKafka processor, so NiFi sends data as a message to
kafka

### Updated Lab 2 - HBase/Hive/Storm
     - Combined the storm and HBase labs together because both involved
deploying a storm topology, viewing the spout/bolt tuples increasing
and topology visualization
     - I took the best parts of Storm lab and HBase/HDFS lab to combine
them.
     - For the HBase/HDFS/Storm Lab, I removed the Storm Configuration
piece where the user verifies storm properly configured. I still
included the important step where they need to turn on the Storm
service.
     - HBase/Hive created tables before deploying storm topology is
necessary, so the trucking demo does not error on Storm UI
     - Added an automation script to setup the extra modules that are
irrelevant from what is being learned in the lab series. This step is
crucial, so that the Storm Topology works properly and no errors occur.

- I removed the code description sections of each lab. I kept the
appendix A section in lab2 where they are guided where to find the
source code in the ion-truck-streaming project folder.

- I added more content to the summary/conclusion sections of each lab,
so the user will know what they learned.

- added trucking demo with NiFi Integration section. In lab 2:
Hbase/Hive/Storm, Appendix A explains how to run the trucking demo and
include NiFi.

### iot-truck-streaming
     - I modified the maven path at the end of setup_demo.sh script, so
that this automation script works.
     - I modified the maven path at the end of the start_demo.sh
script, so that the trucking demo can be deployed on the web

I needed to modify the maven directory paths because it changes when
you manually install maven as the user does in the IoT Lab Series. If
you keep the previous paths, an error stating “directory not found”
will appear.

Note: On my github repo, I still have the original storm lab just in
case, I missed crucial information that should be included in the
merged lab.

@ZacBlanco
@DhruvKumar
@abajwa-hw